backend/requirements.txt
# LimeStar Backend Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlmodel>=0.0.22
aiosqlite>=0.20.0

# AI Service
openai>=1.50.0

# Web Scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Telegram Bot (Phase 7)
# python-telegram-bot>=21.0
